requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
distro==1.9.0
fastapi==0.115.2
greenlet==3.1.1
grpcio==1.66.2
grpcio-tools==1.66.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jiter==0.6.1
numpy==2.1.2
openai==1.51.2
portalocker==2.10.1
protobuf==5.28.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.12
pywin32==308
qdrant-client==1.12.0
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
